requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pytesseract==0.3.10
Pillow==10.4.0
requests==2.31.0
